+++ battlecode22-scaffold/CHANGELOG.md
@@ -1,2 +1,4 @@
 # FirstPageBot
 * Implement archon fleeing, may tweak value and other stuff regulating when it flees
+* Soldiers will now update comms with enemy archon location
+